Cleanup addons/addoninfo/AddonExtensions.h: Get rid of StringUtils.h include.
