+++ xbmc/addons/addoninfo/AddonExtensions.h
@@ -8,8 +8,6 @@
 
 #pragma once
 
-#include "utils/StringUtils.h"
-
 #include <stdlib.h>
 #include <string>
 #include <vector>
@@ -24,9 +22,9 @@ struct SExtValue
 {
   explicit SExtValue(const std::string& strValue) : str(strValue) { }
   const std::string& asString() const { return str; }
-  bool asBoolean() const { return StringUtils::EqualsNoCase(str, "true"); }
-  int asInteger() const { return atoi(str.c_str()); }
-  float asFloat() const { return static_cast<float>(atof(str.c_str())); }
+  bool asBoolean() const;
+  int asInteger() const { return std::atoi(str.c_str()); }
+  float asFloat() const { return static_cast<float>(std::atof(str.c_str())); }
   bool empty() const { return str.empty(); }
   const std::string str;
 };
