+++ xbmc/cores/playercorefactory/PlayerCoreConfig.h
@@ -8,18 +8,12 @@
 
 #pragma once
 
-#include "cores/ExternalPlayer/ExternalPlayer.h"
-#include "cores/IPlayer.h"
-#include "cores/RetroPlayer/RetroPlayer.h"
-#include "cores/VideoPlayer/VideoPlayer.h"
-#include "cores/paplayer/PAPlayer.h"
-#ifdef HAS_UPNP
-#include "network/upnp/UPnPPlayer.h"
-#endif
-#include "utils/XBMCTinyXML.h"
-#include "utils/log.h"
+#include <memory>
+#include <string>
 
-#include <utility>
+class IPlayer;
+class IPlayerCallback;
+class TiXmlElement;
 
 class CPlayerCoreConfig
 {
@@ -27,23 +21,7 @@ public:
   CPlayerCoreConfig(std::string name,
                     std::string type,
                     const TiXmlElement* pConfig,
-                    const std::string& id = "")
-    : m_name(std::move(name)), m_id(id), m_type(std::move(type))
-  {
-    m_bPlaysAudio = false;
-    m_bPlaysVideo = false;
-
-    if (pConfig)
-    {
-      m_config.reset(static_cast<TiXmlElement*>(pConfig->Clone()));
-      const char *sAudio = pConfig->Attribute("audio");
-      const char *sVideo = pConfig->Attribute("video");
-      m_bPlaysAudio = sAudio && StringUtils::CompareNoCase(sAudio, "true") == 0;
-      m_bPlaysVideo = sVideo && StringUtils::CompareNoCase(sVideo, "true") == 0;
-    }
-
-    CLog::Log(LOGDEBUG, "CPlayerCoreConfig::<ctor>: created player {}", m_name);
-  }
+                    const std::string& id = "");
 
   ~CPlayerCoreConfig() = default;
 
@@ -67,52 +45,12 @@ public:
     return m_bPlaysVideo;
   }
 
-  std::shared_ptr<IPlayer> CreatePlayer(IPlayerCallback& callback) const
-  {
-    std::shared_ptr<IPlayer> player;
-
-    if (m_type.compare("video") == 0)
-    {
-      player = std::make_shared<CVideoPlayer>(callback);
-    }
-    else if (m_type.compare("music") == 0)
-    {
-      player = std::make_shared<PAPlayer>(callback);
-    }
-    else if (m_type.compare("game") == 0)
-    {
-      player = std::make_shared<KODI::RETRO::CRetroPlayer>(callback);
-    }
-    else if (m_type.compare("external") == 0)
-    {
-      player = std::make_shared<CExternalPlayer>(callback);
-    }
-
-#if defined(HAS_UPNP)
-    else if (m_type.compare("remote") == 0)
-    {
-      player = std::make_shared<UPNP::CUPnPPlayer>(callback, m_id.c_str());
-    }
-#endif
-    else
-      return nullptr;
-
-    if (!player)
-      return nullptr;
-
-    player->m_name = m_name;
-    player->m_type = m_type;
-
-    if (player->Initialize(m_config.get()))
-      return player;
-
-    return nullptr;
-  }
+  std::shared_ptr<IPlayer> CreatePlayer(IPlayerCallback& callback) const;
 
   std::string m_name;
   std::string m_id; // uuid for upnp
   std::string m_type;
-  bool m_bPlaysAudio;
-  bool m_bPlaysVideo;
+  bool m_bPlaysAudio{false};
+  bool m_bPlaysVideo{false};
   std::unique_ptr<TiXmlElement> m_config;
 };
