+++ xbmc/playlists/PlayListXSPF.cpp
@@ -11,6 +11,7 @@
 
 #include "FileItem.h"
 #include "URL.h"
+#include "utils/StringUtils.h"
 #include "utils/URIUtils.h"
 #include "utils/XBMCTinyXML.h"
 #include "utils/log.h"
